+++ src/main/java/io/github/imsejin/common/assertion/util/CollectionAssert.java
@@ -91,7 +91,7 @@ public class CollectionAssert<
      * @return this class
      */
     @Override
-    public SELF hasSize(Integer expected) {
+    public SELF hasSize(long expected) {
         if (actual.size() != expected) {
             setDefaultDescription(SizeAssertable.DEFAULT_DESCRIPTION_HAS_SIZE, expected, actual.size());
             throw getException();
@@ -107,7 +107,7 @@ public class CollectionAssert<
      * @return this class
      */
     @Override
-    public SELF doesNotHaveSize(Integer expected) {
+    public SELF doesNotHaveSize(long expected) {
         if (actual.size() == expected) {
             setDefaultDescription(SizeAssertable.DEFAULT_DESCRIPTION_DOES_NOT_HAVE_SIZE, expected, actual.size());
             throw getException();
